requirements.txt
streamlit>=1.18.0
transformers>=4.36.2
torch>=1.10.0
sentencepiece
speechrecognition
pydub
gTTS
streamlit-audiorecorder
pyttsx3

